Rearranging overcharge macros
