+++ src/module/macros.ts
@@ -48,6 +48,7 @@ import { prepareTextMacro, rollTextMacro } from "./macros/text"
 export { encodeMacroData, runEncodedMacro } from "./macros/util"
 export { renderMacroTemplate, renderMacroHTML } from "./macros/render"
 export { prepareActivationMacro } from "./macros/activation"
+export { prepareOverchargeMacro } from "./macros/overcharge"
 export { prepareTechMacro } from "./macros/tech"
 export { prepareTextMacro } from "./macros/text"
 
@@ -1058,63 +1059,6 @@ export async function prepareCorePassiveMacro(a: string) {
   };
 
   rollTextMacro(mech, mData).then();
-}
-
-export async function prepareOverchargeMacro(a: string) {
-  // Determine which Actor to speak as
-  let actor = getMacroSpeaker(a);
-  if (!actor) return;
-
-  // Validate that we're overcharging a mech
-  if (!actor.is_mech()) {
-    ui.notifications!.warn(`Only mechs can overcharge!`);
-    return;
-  }
-
-  // And here too... we should probably revisit our type definitions...
-  let rollText = actor.getOverchargeRoll();
-  if (!rollText) {
-    ui.notifications!.warn(`Error in getting overcharge roll...`);
-    return;
-  }
-
-  // Prep data
-  let roll = await new Roll(rollText).evaluate({ async: true });
-
-  let mech = actor.data.data.derived.mm!;
-
-  let mData: LancerOverchargeMacroData = {
-    level: mech.OverchargeCount,
-    roll: roll,
-  };
-
-  // Assume we can always increment overcharge here...
-  mech.OverchargeCount = Math.min(mech.OverchargeCount + 1, 3);
-
-  // Only increase heat if we haven't disabled it
-  if (getAutomationOptions().overcharge_heat) {
-    mech.CurrentHeat = mech.CurrentHeat + roll.total!;
-  }
-
-  await mech.writeback();
-
-  return rollOverchargeMacro(actor, mData);
-}
-
-async function rollOverchargeMacro(actor: LancerActor, data: LancerOverchargeMacroData) {
-  if (!actor) return Promise.resolve();
-
-  const roll_tt = await data.roll.getTooltip();
-
-  // Construct the template
-  const templateData = {
-    actorName: actor.name,
-    roll: data.roll,
-    level: data.level,
-    roll_tooltip: roll_tt,
-  };
-  const template = `systems/${game.system.id}/templates/chat/overcharge-card.hbs`;
-  return renderMacroTemplate(actor, template, templateData);
 }
 
 export function prepareStructureSecondaryRollMacro(registryId: string) {
